apps/api/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
pydantic>=2.7.4,<3.0.0
python-dotenv>=1.0.1

# LangChain — all ecosystem packages pinned together as a compatible set.
# Pinning the whole group prevents pip's backtracking algorithm from choking
# on circular version constraints between langchain/langchain-core/langchain-community.
langchain==0.3.28
langchain-core==0.3.84
langchain-groq==0.2.5
langchain-google-genai==2.1.12
langchain-community==0.3.31
langsmith>=0.3.45,<1.0.0

# Vector + embeddings
pgvector==0.2.5
sentence-transformers==2.7.0
supabase==2.4.3

# L1 inference (ONNX models + SHAP computation)
onnxruntime>=1.18.0,<2.0.0
numpy>=1.26.4,<2.0.0

# Utilities
pyyaml>=6.0.1
httpx>=0.27.0,<1.0.0
python-frontmatter==1.1.0
